Настроил логику авто удаления кода после подтверждения кода

+++ utils/deleteCurrentOtpCode.js
@@ -2,7 +2,9 @@ import OtpCode from '../model/OtpCodeModel.js';
 
 const deleteCurrentOtpCode = async (id, seconds) => {
   const otpCode = await OtpCode.findOne({ id });
-  setTimeout(() => otpCode.deleteOne(), seconds * 1000);
+  if (otpCode) {
+    setTimeout(() => otpCode.deleteOne(), seconds * 1000);
+  }
 };
 
 export default deleteCurrentOtpCode;
